% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_2.R
\name{suma_2}
\alias{suma_2}
\title{suma_2}
\usage{
suma_2(a, b)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}
}
\value{
la suma de a y b
}
\description{
Suma 2 numeros
}
\examples{
suma_2(5,4)
suma_2(-7,18)
}
