% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_3.R
\name{suma_3}
\alias{suma_3}
\title{suma_3}
\usage{
suma_3(a, b, c)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}

\item{c}{tercer numero}
}
\value{
la suma de a, b y c
}
\description{
Suma 3 numeros
}
\examples{
suma_3(5,4,7)
suma_3(8,-4,7)
}
